separate the client API and split the file into components using Routing (React Router)
