+++ package.json
@@ -16,6 +16,7 @@
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-leaflet": "^5.0.0",
+    "react-router-dom": "^7.8.2",
     "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
