+++ src/components/auth/Login.jsx
@@ -1,0 +1,97 @@
+// src/components/auth/Login.jsx
+import React, { useState } from 'react';
+import { useAuth } from '../../context/AuthContext';
+import { API_CONFIG } from '../../api/apiClient';
+
+const Login = () => {
+  const { login } = useAuth();
+  const [formData, setFormData] = useState({ username: '', password: '' });
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
+
+  const handleChange = (e) => {
+    setFormData({
+      ...formData,
+      [e.target.name]: e.target.value
+    });
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError('');
+    setLoading(true);
+
+    const result = await login(formData);
+    
+    if (!result.success) {
+      setError(result.message || 'Login failed');
+    }
+    
+    setLoading(false);
+  };
+
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-600 to-blue-800 flex items-center justify-center">
+      <div className="bg-white rounded-lg shadow-xl p-8 w-full max-w-md">
+        <div className="text-center mb-8">
+          <h1 className="text-3xl font-bold text-gray-800 mb-2">Fleet Monitor</h1>
+          <p className="text-gray-600">Sistem Monitoring Truk Tambang</p>
+          <p className="text-sm text-blue-600 mt-2">PT Borneo Indobara</p>
+        </div>
+        
+        <form onSubmit={handleSubmit}>
+          <div className="mb-4">
+            <label className="block text-gray-700 text-sm font-bold mb-2">
+              Username
+            </label>
+            <input
+              type="text"
+              name="username"
+              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+              value={formData.username}
+              onChange={handleChange}
+              placeholder="Masukkan username"
+              required
+            />
+          </div>
+          
+          <div className="mb-6">
+            <label className="block text-gray-700 text-sm font-bold mb-2">
+              Password
+            </label>
+            <input
+              type="password"
+              name="password"
+              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+              value={formData.password}
+              onChange={handleChange}
+              placeholder="Masukkan password"
+              required
+            />
+          </div>
+          
+          {error && (
+            <div className="mb-4 p-3 bg-red-100 border border-red-400 text-red-700 rounded">
+              {error}
+            </div>
+          )}
+          
+          <button
+            type="submit"
+            disabled={loading}
+            className="w-full bg-blue-600 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 disabled:opacity-50"
+          >
+            {loading ? 'Connecting...' : 'Login'}
+          </button>
+        </form>
+        
+        <div className="mt-4 text-center text-sm text-gray-600">
+          <p>Demo: username: <code>admin</code>, password: <code>admin123</code></p>
+          <p className="mt-2">Backend: {API_CONFIG.BASE_URL}</p>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Login;
